Revert "refactor: twitter api (#8011)"
This reverts commit ff19f897a3c629f7107bebbdabcf91b118e60126.
