+++ packages/mask/src/social-network-adaptor/twitter.com/customization/custom.ts
@@ -1,10 +1,10 @@
-import { useMemo } from 'react'
-import produce, { setAutoFreeze } from 'immer'
 import { MutationObserverWatcher, ValueRef } from '@dimensiondev/holoflows-kit'
 import { createSubscriptionFromValueRef } from '@masknet/shared-base'
-import type { SocialNetworkUI } from '@masknet/types'
 import { useValueRef } from '@masknet/shared-base-ui'
 import { PaletteMode, Theme, unstable_createMuiStrictModeTheme } from '@mui/material'
+import produce, { setAutoFreeze } from 'immer'
+import { useMemo } from 'react'
+import type { SocialNetworkUI } from '@masknet/types'
 import { fromRGB, getBackgroundColor, getForegroundColor, isDark, shade, toRGB } from '../../../utils/theme/index.js'
 import { isMobileTwitter } from '../utils/isMobile.js'
 import { composeAnchorSelector, composeAnchorTextSelector, headingTextSelector } from '../utils/selector.js'
